StackMap is crazy and I love it.
It's a lightly modified version of SquidLib's OrderedMap, which itself
is a modified version of fastutil's Object2ObjectLinkedOpenHashMap
(which retroactively became closer to fastutil's
Object2ObjectLinkedOpenCustomHashMap when the custom IHasher option was
added). It's used here to implement lexical scoping, where a key can be
added many times with different values and older values are kept around
but aren't accessible until newer ones leave scope. This is probably the
strangest data structure I have ever implemented.