Implement = as the assignment macro
The = macro is almost exactly like def, but returns the value it
assigned. It is a macro because it needs to not evaluate the name given
to it, since it might not have a value.

This should fit into a larger pattern of reusing operator names wherever
possible from Java. The one operator we can't use is the unary bitwise
negation operator, `~`, because the tilde is reserved for comments, but
`!^` can substitute in for it with a fairly clear meaning (`!` as "not",
`^` as in "bitwise XOR", where bitwise XOR used on one argument would be
the same as that argument, and "not this argument" is essentially what
bitwise-negate does to every bit). A nice quirk of the syntax we have
here is that we can have a NOT-AND, AND-NOT, NOT-OR, and so on function
constructed by stacking the function names together, like !& for NOT-AND
or &! for AND-NOT, which is difficult in Java-like languages and even
C++ with operator overloading, but easy here with Lisp-style syntax.
Getting the core functions to overload on types may be a bit of a
hassle, but once in the language these should all be straightforward to
define. It also makes it easy to make plain-English names for these
functions, like `and` for boolean-logic `&&` or `and-not` for what is `x
&& !y` in Java, which is probably preferable outside of some specific or
common cases. Symbols could be used for bitwise code as a convention
that fits the already-cryptic-and-terse nature of such code, while words
could be used for the more common logic functions, which also helps
avoid accidental usage of bitwise functions.