Add "and", "or", and bitwise shift functions
This also fixes a bug with hex long parsing (it would always parse a
long in hexadecimal as two tokens, one of which was definitely
incorrect). The macros "and" and "or" are logical (treating arguments as
boolean), and short-circuit-evaluate when given two arguments and the
first satisfies the condition. They don't work for more or less than 2
forms as arguments, currently; they should soon. Bitwise shifts are as
per Java; "<<", ">>", and ">>>" all are implemented as functions and
only do their expected behavior when given 2 arguments (all return 0
when given different argument counts).