Fix bug with repeated calls to a fn; add "defn"
The "def" and "defn" macros also go back to using Context.push instead
of Context.set, which should help the case currently tested in
TokenizerTest. This is where an earlier value used by the Context of a
defined fn is later redefined, but due to scoping the defined fn
preserves the original meaning of its code when it was defined.