Add defmacro, various nice improvements
ARun no longer creates a temporary List containing the function body
just to repopulate the body after modifications; the actual body is now
a final List field, bodyFixed, that can be used to repopulate the
"working" body after it changes due to running. A concrete class, Macro,
acts like ARun but doesn't need a unique implementation per instance,
and can simply be given a List of Token values with bounds on where the
names and body are located. Macro is used to implement defmacro; macro
was planned, by analogy to defmacro as fn is to defn, but issues with
the evaluation have either postponed this or made it impractical, so
defmacro is all there is for now. Context and Token's hashCode method
default to using SquidLib's new Wisp hash now, since it should be faster
and have good quality.