First looping macro: 'while'
It acts mostly as you'd expect, except that it can produce multiple
values, or none, depending on the body. It simply runs the body until
the condition becomes false, accumulating the returned values and
returning all of them. Soon there should be better constructs for
dealing with multiple arguments/returns.