Use non-generic TList for Token lists, add #set[]
TList has some specialized behavior for its one element type, and can
use its awareness that it holds only Tokens to get the solid parts out
of those Tokens in the as() method. This class is returned by the
tokenizing methods and Context.solidify(), which gets only the solid
parts of the results of evaluating a TList. This makes
Context.evaluate() trivial; it only needs to call TList.as() .

The #set[] mode produces an OrderedSet<Token> currently, but will
probably be changed to return a subclass of that which is specialized
for use with Tokens (like TList).