Basic method calls (only for TSet now) and #0()
These two features complement each other. Method calls work without
using reflection, so they should be OK on GWT. the #0( ... ) mode is a
normal function (or macro) call that resolves its arguments like () or
#() but always returns zero values (like calling a void method). This
can be handy with things like the add method on Sets, which returns a
boolean... but we usually don't want it (sometimes we do, but not very
often).