Make seq Iterators in Table usable, fix bug
Removing and re-adding a sequential Integer key wouldn't set the
sequentialLimit to the correct value, only to the most-recently-added
Integer key. If higher keys were present, even if they just became part
of the sequence, the limit wasn't set; now it is. Name has some small
changes, but it might not get used at all.